Add sanitize_dates to utils and use in scrapers
