+++ python/rubin/rag/scrapers/scrape_lsst_bib.py
@@ -38,6 +38,7 @@ from scrapers.utils import (
     batch_by_tokens,
     chunk_docs,
     load_progress,
+    sanitize_dates,
     save_progress,
     write_batches_to_pickle,
 )
@@ -141,7 +142,8 @@ def try_scrape_pdf(
         if not docs or len(docs) < 2:
             _log.warning(f"Small document from PDF at {full_url}")
 
-        chunked = chunk_docs(docs)
+        sanitized = sanitize_dates(docs)
+        chunked = chunk_docs(sanitized)
         batched = batch_by_tokens(chunked)
         write_batches_to_pickle(batched, paper_name, output_dir)
 
@@ -179,7 +181,8 @@ def try_scrape_web(url: str, paper_name: str, output_dir: Path) -> bool:
         if not docs or len(docs[0].page_content) < 2000:
             _log.warning(f"Small document from webpage at {url}")
 
-        chunked = chunk_docs(docs)
+        sanitized = sanitize_dates(docs)
+        chunked = chunk_docs(sanitized)
         batched = batch_by_tokens(chunked)
         write_batches_to_pickle(batched, paper_name, output_dir)
